RF: move test into its own test_ file so gets ran
